+++ CHANGELOG.md
@@ -13,6 +13,8 @@ all failures and stop emitting `UserWarning`.
 ### New Features
 
 - Add `#{state}wx` hash tags to NWS Local Storm Report social media posts.
+- Introduce `pyiem.webutil.iemapp` decorator to do fancy things for IEM
+mod_wsgi apps and remove downstream boilerplate.
 
 ### Bug Fixes